app/requirements.txt
requests
pandas
torch
langsmith
langchain-huggingface
langchain-text-splitters
langchain-chroma
langchain-core
pandas
streamlit
pytest
pytest-asyncio
fastapi
python-dotenv